Time: 14 ms (19.69%), Space: 22.9 MB (77.62%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -3,10 +3,10 @@ public:
     int minEatingSpeed(vector<int>& piles, int h) {
         int low = 1,high = *max_element(piles.begin(),piles.end());
 
-        int reqtime = INT_MAX;
+        long long reqtime = INT_MAX;
         while(low <= high)
         {
-            int mid = low + (high - low) / 2;
+            long long mid = low + (high - low) / 2;
             if(reqTime(piles,mid,h))
             {
                 reqtime = min(reqtime,mid);
@@ -23,7 +23,7 @@ public:
 
     bool reqTime(vector<int>& piles,int hourly,int deadline)
     {
-        int totalhours = 0;
+        long long totalhours = 0;
 
         for(int i = 0;i < piles.size();i++)
         {
